requirements.txt
RPi.GPIO==0.7.0
tinydb==3.15.2
